Move all classes in NewNetworking.Shared to NewNetworking
