+++ Assets/Exanite.Arpg/NewNetworking/Client/UnityClient.cs
@@ -3,7 +3,6 @@ using System.Net;
 using System.Net.Sockets;
 using Exanite.Arpg.Logging;
 using Exanite.Arpg.Networking.Client;
-using Exanite.Arpg.NewNetworking.Shared;
 using LiteNetLib;
 using LiteNetLib.Utils;
 using UniRx.Async;
