+++ Assets/Exanite.Arpg/NewNetworking/EventHandler.cs
@@ -1,4 +1,4 @@
-﻿namespace Exanite.Arpg.NewNetworking.Shared
+﻿namespace Exanite.Arpg.NewNetworking
 {
     public delegate void EventHandler<TSender, TEventArgs>(TSender sender, TEventArgs args);
 }
