--- Assets/Exanite.Arpg/NewNetworking/Shared.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 1bd604e4db03d52499ccbc042dddaecd
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
